+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+Sept 4, 2024
+1. Python virtual env by default.  This means python3-venv is now a hard requirement.
+2. for post chown use logname; resolves sudo su and not invoking with sudo
+3. roll playx3d
+
 Aug 17, 2024
 1. maplibre - enable vulkan backend
 2. move libcamera to master